requirements.txt
PyQt5>=5.15.0
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
cairosvg>=2.5.0
pyinstaller>=4.5.0